Request persistent storage on the client (#11)
This PR adds a request for persistent storage for cloud-v3. This make it
so that clients can have long-lived localStorage and cookies. This is
useful because we want to install the website on users' devices as an
app.

This works in Firefox, but not Chrome or Safari. On this page it says
Chrome automatically handles this request:
https://web.dev/articles/persistent-storage . Perhaps Safari does the
same.